// frontend/src/api/generated/models/NavigationItemComponent.ts
/* tslint:disable */
/* eslint-disable */
/**
 * DOCUMENTATION
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: contact-email@something.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { NavigationSubItemComponent } from './NavigationSubItemComponent';
import {
    NavigationSubItemComponentFromJSON,
    NavigationSubItemComponentFromJSONTyped,
    NavigationSubItemComponentToJSON,
    NavigationSubItemComponentToJSONTyped,
} from './NavigationSubItemComponent';

/**
 * 
 * @export
 * @interface NavigationItemComponent
 */
export interface NavigationItemComponent {
    /**
     * 
     * @type {number}
     * @memberof NavigationItemComponent
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof NavigationItemComponent
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof NavigationItemComponent
     */
    link?: string;
    /**
     * 
     * @type {Array<NavigationSubItemComponent>}
     * @memberof NavigationItemComponent
     */
    subItems?: Array<NavigationSubItemComponent>;
}

/**
 * Check if a given object implements the NavigationItemComponent interface.
 */
export function instanceOfNavigationItemComponent(value: object): value is NavigationItemComponent {
    return true;
}

export function NavigationItemComponentFromJSON(json: any): NavigationItemComponent {
    return NavigationItemComponentFromJSONTyped(json, false);
}

export function NavigationItemComponentFromJSONTyped(json: any, ignoreDiscriminator: boolean): NavigationItemComponent {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'link': json['link'] == null ? undefined : json['link'],
        'subItems': json['subItems'] == null ? undefined : ((json['subItems'] as Array<any>).map(NavigationSubItemComponentFromJSON)),
    };
}

export function NavigationItemComponentToJSON(json: any): NavigationItemComponent {
    return NavigationItemComponentToJSONTyped(json, false);
}

export function NavigationItemComponentToJSONTyped(value?: NavigationItemComponent | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'link': value['link'],
        'subItems': value['subItems'] == null ? undefined : ((value['subItems'] as Array<any>).map(NavigationSubItemComponentToJSON)),
    };
}